Removed TraviCI
When I submitted meantables to CRAN, Travis was causing an error.

Found the following (possibly) invalid URLs:
  URL: https://travis-ci.com/brad-cannell/meantables (moved to https://www.travis-ci.com/brad-cannell/meantables)
    From: README.md
    Status: 404
    Message: Not Found

I tried to fix the error on Travis, but I'm not sure what's going on. Travis really isn't necessary for a package this small anyway. I decided to just go ahead and remove it for now. I can always add it again later if I need to.